Update patch set 2
Patch Set 2:

This need not be done this way..
The change of the individual pom files needs to be done separately later.
For now  I have bumoed the version of SO that will take care of the 1.7.6 release version for the entire so pods.

Patch-set: 2
Reviewer: Gerrit User 897 <897@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>