Update patch set 1
Patch Set 1: Code-Review-2

What is the reason of that change?

The value skipPostInstConf was recently changed to primitive because of this bug: https://jira.onap.org/browse/SO-3346
Shortly, when this value is non-primitive (Boolean object), there is a problem with getting the value from mariaDB.
The same issue is in case of pnf.

Beside, what is the advantage of using object instead of primitive in this case?

Patch-set: 1
Reviewer: Gerrit User 2789 <2789@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=-2