Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-2
> 
> What is the reason of that change?
> 
> The value skipPostInstConf was recently changed to primitive because of this bug: https://jira.onap.org/browse/SO-3346
> Shortly, when this value is non-primitive (Boolean object), there is a problem with getting the value from mariaDB.
> The same issue is in case of pnf.
> 
> Beside, what is the advantage of using object instead of primitive in this case?

Database entities should always use non primitives. This should not cause any issue in the future because we're are explicitly setting a default within the Java class this time. The root cause of this issue is that the DB scripts were written to define a default but not "NON NULL" which allows null values to be written to rows in the DB. ASDC Controller is explicitly writing a null for this value however, setting true within the java class will handle existing and future cases.

If there is a test that should be run before or after this is merged. This can be done.

Patch-set: 1