Update patch set 1
Patch Set 1: -Code-Review

> Patch Set 1:
> 
> > Patch Set 1: Code-Review-2
> > 
> > What is the reason of that change?
> > 
> > The value skipPostInstConf was recently changed to primitive because of this bug: https://jira.onap.org/browse/SO-3346
> > Shortly, when this value is non-primitive (Boolean object), there is a problem with getting the value from mariaDB.
> > The same issue is in case of pnf.
> > 
> > Beside, what is the advantage of using object instead of primitive in this case?
> 
> Database entities should always use non primitives. This should not cause any issue in the future because we're are explicitly setting a default within the Java class this time. The root cause of this issue is that the DB scripts were written to define a default but not "NON NULL" which allows null values to be written to rows in the DB. ASDC Controller is explicitly writing a null for this value however, setting true within the java class will handle existing and future cases.
> 
> If there is a test that should be run before or after this is merged. This can be done.

The skipPostInstConf value is not a entity, only a member of entity. But as you wrote, if a column in database allows null value, then Boolean object is a proper option.

But I have another concerns. In case of pnf, we had a situation, that even if the value in database was not null, it could not be assigned to Boolean object, and the result was null pointer exception. After changing to primitive, the problem disappeared.

I see you changed the way of getting the value, by adding:
vrc.isSkipPostInstConf();  ->  vrc.isSkipPostInstConf().booleanValue();

Maybe it can help, have you tested it on the lab? I am not sure if there is test that can handle this situation.
Anyway, I remove "-2" not to block this ticket, but I would prefer to merge it after confirmation, that it was tested and the value is got from database properly.

Patch-set: 1
Reviewer: Gerrit User 2789 <2789@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0