Revert "Changing docker module packaging to JAR"
This reverts commit 40fc9d52b0ca2eaba6d1e30f846797e301dc885b.

Reason for revert: it seems that docker module in SO is triggered by “docker” profile which is only triggered in so-maven-docker-* Jenkins job. Per my understanding these jobs are only used for creating docker images and I think relying on these jobs to create a docker JAR module will create unnecessary dependency. So I think option 3 is more feasible

Change-Id: Ia64778afd8fe4105ee8bf1a2f45a4d5674601289
Issue-ID: SO-3473
Signed-off-by: waqas.ikram <waqas.ikram@est.tech>