Update patch set 2
Patch Set 2:

Even though the merge job worked before the code review was merged. It does not seem like works after it merged. The so-settings.xml file needs to be updated probably. I could not find this file in any repository. Is it located on the jenkins instance itself?

Patch-set: 2