Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > > Patch Set 2:
> > > 
> > > > Patch Set 2: Code-Review-1
> > > > 
> > > > As per SECCOM/OWASP/<Name security authority of your choice> recommendation you should not hardcode credentials in your source code
> > > 
> > > there are just dummy values to unblock oom
> > 
> > NO. Those are not dummy values those are hardcoded credentials that should not be there.
> > Please don't violate SECCOM recommendations and follow best practices established by security authorities:
> > 
> > https://www.beyondtrust.com/resources/glossary/hardcoded-embedded-passwords
> > https://owasp.org/www-community/vulnerabilities/Use_of_hard-coded_password
> > https://cwe.mitre.org/data/definitions/259.html
> > 
> > There is no option that we'll merge your updated image in OOM with such hardcoded passwords. As soon as it arrives in the review you will get -2 from my side.
> 
> We're happy to fix it based on these standards and are working to do so in https://gerrit.onap.org/r/c/oom/+/119454.  This a completely non-sense password that we're just providing as a way to temporarily unblock this issue.

Thank you for your cooperation and I believe that you are trying to achieve the right target but just chosen the wrong path to achieve it.
Please use common secret template and init container as so-cnf-adapter is doing:
https://github.com/onap/oom/blob/master/kubernetes/so/components/so-cnf-adapter/values.yaml
instead of continuing with this patch.

Actually the more non-sense password is the more secure it gets right? Perfectly random strings tends to be the best passwords;)

I know that this password is compromised and published in the web but please don't increase number of places where it is hardcoded by adding this patch.

Patch-set: 2
Reviewer: Gerrit User 4965 <4965@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>