Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > > Patch Set 2: Code-Review-1
> > > 
> > > As per SECCOM/OWASP/<Name security authority of your choice> recommendation you should not hardcode credentials in your source code
> > 
> > there are just dummy values to unblock oom
> 
> NO. Those are not dummy values those are hardcoded credentials that should not be there.
> Please don't violate SECCOM recommendations and follow best practices established by security authorities:
> 
> https://www.beyondtrust.com/resources/glossary/hardcoded-embedded-passwords
> https://owasp.org/www-community/vulnerabilities/Use_of_hard-coded_password
> https://cwe.mitre.org/data/definitions/259.html
> 
> There is no option that we'll merge your updated image in OOM with such hardcoded passwords. As soon as it arrives in the review you will get -2 from my side.

Actually they are dummy values. How do figure they are not? If you decrypt this hardcoded default value then you will see that it literally evaluates to a value equal too "defaultAuth".  I call that a dummy value. None the less we have only made this change to try to help oom because they cant figure out how to add the configuration/properties to their helm even though I clearly outlined in an email which properties needed to be added.

Patch-set: 2