Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 1: Code-Review-1
> > 
> > (1 comment)
> > 
> > I’m not sure hard coding credentials is safe...
> 
> I agree. These are not real credentials, nor a real key. They are just dummy default values. If you decrypt them you will see the evaluate to a value of "defaultAuth"

so why not let the code "crash" if the value is not provided? 
it would be a lot safer as it would enforce users to provide a value instead of using this dummy value if they don't know they have to provide it no?

Patch-set: 2
Reviewer: Gerrit User 2319 <2319@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>