Enable WorkflowActionBB to skip last building block
WorkflowActionBB does not take into account flowmanipulators such as
SkipCDSBuildingBlockListener, which increment the current index of
building blocks to execute. In the case where ControllerExecutionBB is
the last building block and should be skipped, the index would be
incremented beyond bounds and cause exception. WorkflowActionBB needs
to check if the flow has been completed before it tries to execute the
next building block.

Issue-ID: SO-3678
Change-Id: I635c12a568c3b98031cbeb37ef521663d96b852b
Signed-off-by: Jozsef Csongvai <jozsef.csongvai@bell.ca>