Update patch set 4
Patch Set 4:

I tested this by creating a fresh database first and then modify R__MacroData with the changes in this patch.
All tables were exactly the same in content except for orchestration_flow_reference where the SEQ_NO value is now consistent where in the old script you had gaps and sometimes SEQ_NO starting at 3 instead of 1.

I also updated the macro Activate and Assign to use ControllerExecutionBB instead of ConfigAssignBB and ConfigDeployBB to be consistent with the create and delete macro.

Patch-set: 4