SO changes for Service Intent
This update contains the SO changes for the CCVPN extension to
support Intent Based Networking (REQ-719). The SO is responsible
for providing the Service Intent orchestration, as well as providing
the Service Intent APIs which are consumed by the IBN use-case.

The changes on the SO are technology-agnostic and are relatively
simple. i.e., Technology specific network configurations are carried
out by the SDNC.

Issue-ID: SO-3714
Signed-off-by: hyu2010 <hyu2010b@gmail.com>
Change-Id: Iec9d2fc80adc79faaeeaf811bd0d02c9d7611d61