SO-3720 BuildingBlockRollback lookup table
Issue-ID: SO-3720

BuildingBlockRollback lookup table will indicate the rollback BB to take;
this is the first part for simplifying WorkflowActionBBTasks logic:
instead of doing string matching, we just do a lookup of a block to call
to perform the rollback operation.
In some cases, the Action is set as well - this would be needed for certain
BBs, in that case, there is a corresponding rollbackAction as well.

Change-Id: I072a2ada894cf4672f5a1cdce762605757cb1d14
Signed-off-by: Oleg Mitsura <oleg.mitsura@amdocs.com>