Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> > Patch Set 4: Code-Review-1
> > 
> > it looks that this change has a consequence on the basic_vm_macro test... (2 errors at SO level on the 2 last gate)
> > it can be a regression or simply mean that the test must be adapted to match the change.
> > We need to figure out before any merge
> 
> I checked some of last changes in master branch of so. basic_vm_macro also failed there. Which means the failure is not related to my change. 
> The change I referred: https://gerrit.onap.org/r/c/so/+/120538
> Integration results link: https://logs.onap.org/onap-integration/gating//120538-5/index.html

on recent daily chains, the test is still OK.
The error from yesterday daily master was due to multicloud
a gate from today on SO update for Istanbul (gerrit 123386) (master + SO docker candidates) shows everything in green: https://logs.onap.org/onap-integration/gating//123386-2/index.html
 
so it seems that the changes triggering the issue has not been merged in OOM yet. As we are gating on SO it shall be possible to look at the recent SO gates and detect which patch is leading to the issue.

Patch-set: 4
Reviewer: Gerrit User 1215 <1215@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>