Use resource list for SkipCDSBuildingBlockListener
The listener was using a customizationId from the requestPayload.
Since the requestPayload is slightly different for VNF LCM requests,
this logic was inconsistent and not working properly.

By adding modelCustomizationId and modelVersionId to the Resource
class, the SkipCDSBuildingBlockListener can leverage this to perform
more reliable lookups for customization objects.

This also enables SkipCDSBuildingBlockListener to run for delete
requests. Previously this would have failed as the resourceKey is
instance-id in the delete case and customization-id in for create.
By using the explicit property of the resource object, there will
be no confusion.

Issue-ID: SO-3807
Change-Id: I0922e5e7373990d9b409651cb4365e11d5618558
Signed-off-by: Jozsef Csongvai <jozsef.csongvai@bell.ca>