Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Hi @Sreeja.
> Can I ask how a test class relates to OOM smoke test failures?
> Does this patch change any behaviour in how SO reads from the topics?

Hi @Fiachra,
With our analysis, we found out that in recent patch '@Ignore' was used in this test file, due to which PNF_READY was not received, only PNF_UPDATE was received. The same was observed from your end in OOMSmokeTests. 
Now, with this patch it will receive both events in test scenarios, we have checked this locally too. Anyhow, this fix was needed in SO.
And,I presume, OOMSmokeTests was passing earlier with PNF_READY topic.
Now, we will have to check OOMSmokeTests will pass both PNF_READY & PNF_UPDATE, as this patch-set is merged.

Patch-set: 2
Reviewer: Gerrit User 7813 <7813@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>